Add files via upload
Create a robust, bulletproof C# console application that implements the following functionality:

Prompt the user to enter three integers one at a time, then use a method to calculate and return their sum.  Have the Main() method display the result.
.
Prompt the user to enter an integer, then use a method to calculate and return the value of the polynomial 4x3 + 5x – 2 (where x is the number entered by the user).  Have the Main() method display the result.
.
Prompt the user to enter a number of seconds and store their input as an unsigned integer.  Pass their input into a method that converts it into the equivalent number of hours, minutes, and seconds and returns that information as a string.  Have the Main() method display the results for the user.
For example, if the user entered 52400, the program would display "14 hours, 33 minutes, 20 seconds".
.
Prompt the user to enter a sequence of integers one-at-a-time.  Have them enter "end" when they are finished.  After they are finished entering numbers, display the maximum and minimum values from their sequence.
.
Use a while-loop to display a list of all even integers between 0 and 200 (inclusive) in ascending order.  Your output should display 10 numbers per line separated by commas.  There should not be a comma after the very last number in the list.
.
Use a do/while-loop to display a list of all odd integers between 200 and 0 (inclusive) in descending order.  Your output should display 25 numbers per line separated by commas.  There should not be a comma after the very last number in the list.